+++ packages/lib/src/app/App.js
@@ -2,9 +2,9 @@
 
 import Logger from '../utils/Logger'
 import Contracts from '../utils/Contracts'
-import { deploy, sendTransaction } from '../utils/Transactions'
 import decodeLogs from '../helpers/decodeLogs'
 import encodeCall from '../helpers/encodeCall'
+import { sendTransaction } from '../utils/Transactions'
 
 import AppProvider from './AppProvider'
 import AppDeployer from './AppDeployer'
@@ -34,21 +34,16 @@ export default class App {
     this.factory = factory
     this.package = _package
     this.version = version
-    this.directories = {}
-    this.directories[version] = appDirectory
+    this.directory = appDirectory
     this.txParams = txParams
   }
 
-  address() {
+  get address() {
     return this._app.address
   }
 
-  currentDirectory() {
-    return this.directories[this.version]
-  }
-
   async currentStdlib() {
-    return this.currentDirectory().stdlib()
+    return this.directory.stdlib()
   }
 
   async hasStdlib() {
@@ -56,8 +51,7 @@ export default class App {
   }
 
   async getImplementation(contractName) {
-    const directory = this.currentDirectory()
-    return directory.getImplementation(contractName)
+    return this._app.getImplementation(contractName)
   }
 
   async getProxyImplementation(proxyAddress) {
@@ -65,37 +59,23 @@ export default class App {
   }
 
   async setImplementation(contractClass, contractName) {
-    log.info(`Setting implementation of ${contractName} in directory...`)
-    const implementation = await deploy(contractClass, [], this.txParams)
-    const directory = this.currentDirectory()
-    await sendTransaction(directory.setImplementation, [contractName, implementation.address], this.txParams)
-    log.info(` Implementation set: ${implementation.address}`)
-    return implementation
+    return this.package.setImplementation(this.version, contractClass, contractName)
   }
 
   async unsetImplementation(contractName) {
-    log.info(`Unsetting implementation of ${contractName} in directory...`)
-    await sendTransaction(this.currentDirectory().unsetImplementation, [contractName], this.txParams)
-    log.info(`Implementation unset`)
+    await this.directory.unsetImplementation(contractName, this.txParams)
   }
 
   async setStdlib(stdlibAddress = 0x0) {
-    log.info(`Setting stdlib ${stdlibAddress}...`)
-    await sendTransaction(this.currentDirectory().setStdlib, [stdlibAddress], this.txParams)
-    return stdlibAddress
+    return this.directory.setStdlib(stdlibAddress)
   }
 
-  async newVersion(versionName, stdlibAddress = 0) {
-    log.info(`Adding version ${versionName}...`)
-    const AppDirectory = Contracts.getFromLib('AppDirectory')
-    const directory = await deploy(AppDirectory, [stdlibAddress], this.txParams)
-    log.info(` App directory: ${directory.address}`)
-    await sendTransaction(this.package.addVersion, [versionName, directory.address], this.txParams)
-    log.info(` Added version: ${versionName}`)
-    await sendTransaction(this._app.setVersion, [versionName], this.txParams)
-    log.info(` Version set`)
-    this.directories[versionName] = directory
-    this.version = versionName
+  async newVersion(version, stdlibAddress = 0x0) {
+    const directory = await this.package.newVersion(version, stdlibAddress)
+    await sendTransaction(this._app.setVersion, [version], this.txParams)
+    log.info(`Version ${version} set in app.`)
+    this.directory = directory
+    this.version = version
   }
 
   async changeProxyAdmin(proxyAddress, newAdmin) {
@@ -110,11 +90,11 @@ export default class App {
       ? await this._createProxy(contractName)
       : await this._createProxyAndCall(contractClass, contractName, initMethodName, initArgs)
 
-    log.info(` TX receipt received: ${receipt.transactionHash}`)
+    log.info(`TX receipt received: ${receipt.transactionHash}`)
     const UpgradeabilityProxyFactory = Contracts.getFromLib('UpgradeabilityProxyFactory')
     const logs = decodeLogs(receipt.logs, UpgradeabilityProxyFactory)
     const address = logs.find(l => l.event === 'ProxyCreated').args.proxy
-    log.info(` ${contractName} proxy: ${address}`)
+    log.info(`${contractName} proxy: ${address}`)
     return new contractClass(address)
   }
 
@@ -123,7 +103,7 @@ export default class App {
     const { receipt } = typeof(initArgs) === 'undefined'
       ? await this._upgradeProxy(proxyAddress, contractName)
       : await this._upgradeProxyAndCall(proxyAddress, contractClass, contractName, initMethodName, initArgs)
-    log.info(` TX receipt received: ${receipt.transactionHash}`)
+    log.info(`TX receipt received: ${receipt.transactionHash}`)
   }
 
   async _createProxy(contractName) {
@@ -131,11 +111,9 @@ export default class App {
     return sendTransaction(this._app.create, [contractName], this.txParams)
   }
 
-  async _createProxyAndCall(contractClass, contractName, initMethodName, initArgs) {    
-    const initMethod = this._validateInitMethod(contractClass, initMethodName, initArgs)
-    const initArgTypes = initMethod.inputs.map(input => input.type)
+  async _createProxyAndCall(contractClass, contractName, initMethodName, initArgs) {
+    const { initMethod, callData } = this._buildInitCallData(contractClass, initMethodName, initArgs)
     log.info(`Creating ${contractName} proxy and calling ${this._callInfo(initMethod, initArgs)}`)
-    const callData = encodeCall(initMethodName, initArgTypes, initArgs)
     return sendTransaction(this._app.createAndCall, [contractName, callData], this.txParams)
   }
 
@@ -152,9 +130,16 @@ export default class App {
     return sendTransaction(this._app.upgradeAndCall, [proxyAddress, contractName, callData], this.txParams)
   }
 
+  _buildInitCallData(contractClass, initMethodName, initArgs) {
+    const initMethod = this._validateInitMethod(contractClass, initMethodName, initArgs)
+    const initArgTypes = initMethod.inputs.map(input => input.type)
+    const callData = encodeCall(initMethodName, initArgTypes, initArgs)
+    return { initMethod, callData }
+  }
+
   _validateInitMethod(contractClass, initMethodName, initArgs) {
     const initMethod = contractClass.abi.find(fn => fn.name === initMethodName && fn.inputs.length === initArgs.length)
-    if (!initMethod) throw `Could not find initialize method '${initMethodName}' with ${initArgs.length} arguments in contract class`
+    if (!initMethod) throw Error(`Could not find initialize method '${initMethodName}' with ${initArgs.length} arguments in contract class`)
     return initMethod
   }