+++ packages/lib/src/package/Package.js
@@ -1,22 +1,37 @@
 import Logger from '../utils/Logger'
-import Contracts from '../utils/Contracts'
 import { deploy, sendTransaction } from '../utils/Transactions'
 
-import PackageDeployer from './PackageDeployer'
 import PackageProvider from './PackageProvider'
+import PackageDeployer from './PackageDeployer'
 
 const log = new Logger('Package')
 
 export default class Package {
 
-  static async fetch(address, txParams = {}) {
+  static fetch(address, txParams = {}, klass = require('./PackageWithAppDirectories')) {
     const provider = new PackageProvider(txParams)
-    return await provider.from(address)
+    return provider.fetch(address, klass.default)
   }
 
-  static async deploy(txParams = {}) {
+  static async deploy(txParams = {}, klass = require('./PackageWithAppDirectories')) {
     const deployer = new PackageDeployer(txParams)
-    return await deployer.deploy()
+    return deployer.deploy(klass.default)
+  }
+
+  static async deployWithFreezableDirectories(txParams = {}) {
+    return this.deploy(txParams, require('./PackageWithFreezableDirectories'))
+  }
+
+  static fetchWithFreezableDirectories(address, txParams = {}) {
+    return this.fetch(address, txParams, require('./PackageWithFreezableDirectories'))
+  }
+
+  static async deployWithNonFreezableDirectories(txParams = {}) {
+    return this.deploy(txParams, require('./PackageWithNonFreezableDirectories'))
+  }
+
+  static fetchWithNonFreezableDirectories(address, txParams = {}) {
+    return this.fetch(address, txParams, require('./PackageWithNonFreezableDirectories'))
   }
 
   constructor(_package, txParams = {}) {
@@ -24,59 +39,50 @@ export default class Package {
     this.txParams = txParams
   }
 
-  address() {
+  get address() {
     return this.package.address
   }
 
   async hasVersion(version) {
-    return await this.package.hasVersion(version)
-  }
-
-  async getRelease(version) {
-    const releaseAddress = await this.package.getVersion(version)
-    const Release = Contracts.getFromLib('Release')
-    return new Release(releaseAddress)
-  }
-
-  async newVersion(version) {
-    log.info('Adding new version...')
-    const Release = Contracts.getFromLib('Release')
-    const release = await deploy(Release, [], this.txParams)
-    await sendTransaction(this.package.addVersion, [version, release.address], this.txParams)
-    log.info(' Added version:', version)
-    return release
-  }
-
-  async isFrozen(version) {
-    const release = await this.getRelease(version)
-    return await release.frozen()
-  }
-
-  async freeze(version) {
-    log.info('Freezing new version...')
-    const release = await this.getRelease(version)
-    await sendTransaction(release.freeze, [], this.txParams)
-    log.info(' Release frozen')
+    return sendTransaction(this.package.hasVersion, [version], this.txParams)
   }
 
   async getImplementation(version, contractName) {
-    const release = await this.getRelease(version)
-    return await release.getImplementation(contractName)
+    return this.package.getImplementation(version, contractName)
   }
 
   async setImplementation(version, contractClass, contractName) {
     log.info(`Setting implementation of ${contractName} in version ${version}...`)
     const implementation = await deploy(contractClass, [], this.txParams)
-    const release = await this.getRelease(version)
-    await sendTransaction(release.setImplementation, [contractName, implementation.address], this.txParams)
-    log.info(` Implementation set: ${implementation.address}`)
+    const directory = await this.getDirectory(version)
+    await directory.setImplementation(contractName, implementation.address)
     return implementation
   }
 
   async unsetImplementation (version, contractName) {
     log.info(`Unsetting implementation of ${contractName} in version ${version}...`)
-    const release = await this.getRelease(version)
-    await sendTransaction(release.unsetImplementation, [contractName], this.txParams)
-    log.info(`Implementation unset for ${contractName} in version ${version}`)
-  };
+    const directory = await this.getDirectory(version)
+    await directory.unsetImplementation(contractName, this.txParams)
+  }
+
+  async newVersion(version, stdlibAddress) {
+    log.info('Adding new version...')
+    const directory = await this.newDirectory(stdlibAddress)
+    await sendTransaction(this.package.addVersion, [version, directory.address], this.txParams)
+    log.info(`Added version ${version}`)
+    return directory
+  }
+
+  async getDirectory(version) {
+    const directoryAddress = await this.package.getVersion(version)
+    return this.wrapImplementationDirectory(directoryAddress)
+  }
+
+  wrapImplementationDirectory() {
+    throw Error('Cannot call abstract method wrapImplementationDirectory()')
+  }
+
+  async newDirectory() {
+    throw Error('Cannot call abstract method newDirectory()')
+  }
 }